### 2018-09-16 #953 @arfon
@tristan-salles, please make this change to your `paper.md` file:

```
Change this line:

![Example of **eSCAPE** output showing Earth evolution in the future after 100,000 years of constant sea-level and uniform precipitation.](https://github.com/Geodels/eSCAPE/blob/master/images/FA.png)

to this:

![Example of **eSCAPE** output showing Earth evolution in the future after 100,000 years of constant sea-level and uniform precipitation.](https://raw.githubusercontent.com/Geodels/eSCAPE/master/images/FA.png)
```
