### 2018-09-16 2aabb34
Update paper.md
