### 2018-09-16 #953 @tristan-salles
Checked.

Tristan Salles | Lecturer
The University of Sydney
Faculty of Science, School of Geosciences
Rm 454, Madsen Building F09 | The University of Sydney | NSW | 2006
+61 2 8627 4123<tel://8627%204123>  | +61 451 462 502<tel://+61%20451%20462%20502>
tristan.salles@sydney.edu.au<mailto:firstname.surname@sydney.edu.au>  | sydney.edu.au<http://sydney.edu.au/>
[ web-profile<https://tristansalles.github.io/>   |   badlands<http://github.com/badlands-model> ]
CRICOS 00026A
This email plus any attachments to it are confidential. Any unauthorised use is strictly prohibited.
If you receive this email in error, please delete it and any attachments.
Please think of our environment and only print this email if necessary.
