### 2018-09-17 #953 @kthyng
@tristan-salles One high level comment: as per the [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), you don't want any instructions on how to use the code in your paper, so I think your "Dependencies and Examples" section should just be in your code documentation.

Do you have any reviewer suggestions? There is a [list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) available, or perhaps you have someone else in mind.
