### 2018-09-18 #953 @kthyng
@tristan-salles Yes, please go ahead and update your paper accordingly, since otherwise I think the reviewers will request that.

Hi @sgrieve and @kbarnhart! Are you willing to refer this submission for JOSS? You can read about the review process [here](https://joss.readthedocs.io/en/latest/review_criteria.html).
