### 2018-09-18 30cc640
Update paper.md
