### 2018-09-18 #953 @sgrieve
I'd be delighted to review this! I saw some tweets about this and was planning to play with it, and this will make sure I follow through!

It's the start of teaching for me this week, so I won't have a chance till next week, if that's ok.
