### 2018-09-18 #964 @whedon
[REVIEW]: eSCAPE: parallel global-scale landscape evolution model
**Submitting author:** @tristan-salles (<a href="http://orcid.org/0000-0001-6095-7689">Tristan Salles</a>)
**Repository:** <a href="https://github.com/Geodels/eSCAPE" target ="_blank">https://github.com/Geodels/eSCAPE</a>
**Version:** v1.0.0
