### 2018-09-18 #953 @kbarnhart
@kthyng I'd also be happy to review this.
