### 2018-09-18 #964 @kthyng
ok @sgrieve and @kbarnhart I have this ready for you now. Let me know if you have any questions.

@arfon Do you have any guidance for how I can determine why the pdf is not compiling?
