### 2018-09-24 Geodels/eSCAPE#1 @kbarnhart
Cannot find API documentation
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/964) I have looked for API documentation ([review standards listed here](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation)).

I have found an excellent description of the eSCAPE input file [here](https://github.com/Geodels/eSCAPE/wiki/Input-file). And looking at the source code I see that docstrings have been written for the major public-facing functions.

However, I have not found API documentation. As this is one of the review requirements, I'd recommend making more easily visible.
