### 2018-09-24 Geodels/eSCAPE#2 @kbarnhart
eSCAPE needs instructions for testing 
This is another issue associated with [this JOSS review](https://github.com/openjournals/joss-reviews/issues/964).

The [review requirements regarding testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) recommended having either an automated test suite or a set of "manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)".

Given the dependencies of eSCAPE on HPCC-type libraries I can understand why it might be challenging to set up automated tests on standard CI platforms (e.g. Travis).

However, given the large number of (potentially user compiled) dependencies it seems like it would be very good to have one example that could test that eSCAPE produces the expected output. One option would be to provide the `imput_glob.yaml` and `run_espace.py` files discussed [in this part of the wiki](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC#example-of-pbs-script) and add some lines to `run_espace.py` that open the output and assert that some values have been calculated correctly.
