### 2018-09-24 Geodels/eSCAPE#3 @kbarnhart
Improve eSCAPE HPCC installation instructions
This is another issue associated with [this JOSS review](https://github.com/openjournals/joss-reviews/issues/964).

I downloaded and tested the Docker image (and as expected it worked). I wanted to also test the HPCC installation instructions as I would expect that most full-scale applications would use eSCAPE installed in this way.

I'm still in the process of doing this install so I can't yet report if I've been successful. However, thus far I have some comments on the [Install on HPC wiki page](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC).

Specifically, rather than providing what seems to be a set of instructions that have worked on one particular HPC platform, I'd recommend listing the following:

- a list of required compilers
- a statement about what major.minor versions of python eSCAPE is compatible with (2.7. 3.5, 3.6?)
- a package management file that lists the standard python packages (e.g. `numpy`) and minimum version requirements so that these can be installed with `pip` or `conda`.
- a list of the more customized python packages that must be installed by cloning a repo and compiling.
- a statement about the version of PETSc required for running eSCAPE.

Its also a bit confusing to my why the  [installing petsc4py locally](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC#installing-petsc-petsc4py-locally) section is located after the statements about compiling eSCAPE.
