### 2018-09-24 Geodels/eSCAPE#4 @kbarnhart
Improve description of spatially and temporally variable inputs and outputs
This is another issue associated with [the JOSS review](https://github.com/openjournals/joss-reviews/issues/964).

The tutorials have some information about the required format and the way to create the `.vtk` format used in the input files for the domain, the climate, and the uplift, but a description of these inputs is not in the main package documentation.

As best as I can tell there is also no description of the requirements for the sea level curve file.

I'd also recommend including a statement regarding the nature of the output (a file for each timestep that a user can open in paraview OR compile themselves into a single file with a time dimension).
