### 2018-09-24 #964 @kbarnhart
Since I've just loaded on a bunch of issues about documentation, testing, and installation, I thought I'd also make a note here that I think the core functionality of eSCAPE is great, the tutorials are very nicely done, and the docker image works as expected.
