### 2018-09-24 Geodels/eSCAPE#5 @kbarnhart
Are there any mpirun related constraints on number of cores, nodes, etc
And another issue related to [the JOSS review](https://github.com/openjournals/joss-reviews/issues/964).

Are there any constraints, minimums, maximums, etc related to the parallelization of eSCAPE. Or can it use whatever resources a platform can provide and mpirun can specify?
