### 2018-09-24 Geodels/eSCAPE#6 @kbarnhart
Improve contributing/issues/help 
eSCAPE has a slack channel which is great! But based on the JOSS guidelines for [contributing](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) it looks like eSCAPE needs more information regarding contributing and reporting bugs.

[Associated with JOSS Review](https://github.com/openjournals/joss-reviews/issues/964)
