### 2018-09-24 Geodels/eSCAPE#7 @kbarnhart
Incomplete documentation related to boundary conditions
I've tried to limit issues related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/964) to what is outlined in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

There is one aspect of documentation not described in the criteria that I think is missing in eSCAPE. Specifically I can't tell what sort of model-domain edge boundary conditions are used (and if there are any options). This would be a good thing to add to the documentation.
