### 2018-09-25 3e1d50c
add sphinx API documentation
