+++ eSCAPE/mesher/unstructuredmesh.py
@@ -1,20 +1,20 @@
-"""
-Copyright 2017-2018 Tristan Salles
-
-This file is part of eSCAPE.
-
-eSCAPE is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or any later version.
-
-eSCAPE is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with eSCAPE.  If not, see <http://www.gnu.org/licenses/>.
-"""
+###
+# Copyright 2017-2018 Tristan Salles
+#
+# This file is part of eSCAPE.
+#
+# eSCAPE is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or any later version.
+#
+# eSCAPE is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with eSCAPE.  If not, see <http://www.gnu.org/licenses/>.
+###
 
 import numpy as np
 from mpi4py import MPI
@@ -142,6 +142,12 @@ class UnstMesh(object):
         return
 
     def _naturalNumbering(self,coords):
+        """
+        Natural numbering based on DMPlex distribution
+
+        Args:
+            coords: mesh coordinates
+        """
 
         from scipy.spatial import cKDTree as _cKDTree
 
@@ -161,6 +167,12 @@ class UnstMesh(object):
     def _create_DMPlex(self, dim, coords, cells, elev):
         """
         Create a PETSc DMPlex object from the mesh attributes
+
+        Args:
+            dim: mesh dimensions
+            coords: mesh coordinates
+            cells: cell nodes indices
+            elev: nodes elevation
         """
 
         t0 = clock()

[…]
