+++ doc/requirements.txt
@@ -1,4 +1,4 @@
-Sphinx
-mock
 numpy
+Sphinx
+sphinx_rtd_theme
 -e ./
