+++ doc/requirements.txt
@@ -1,5 +1,5 @@
-numpy
 Sphinx
 sphinx_rtd_theme
 setuptools
+numpy
 -e ./
