### 2018-09-25 Geodels/eSCAPE#1 @tristan-salles
I have updated the website and wiki with link to the API documentation:
https://escape-api.github.io/index.html
