### 2018-09-25 Geodels/eSCAPE#2 @tristan-salles
Based on this comment I have provided an additional test in the eSCAPE-demo folder for this purpose:

### Testing eSCAPE installation

The [**eSCAPE-demo**](https://github.com/Geodels/eSCAPE-demo) contains a **test** folder that provides manual steps that can be followed to check the expected functionality of the software.

The **test** proposes tow comparisons related to:

1. the _expected output values_ that you should be obtained if the installation is successful
2. runtime for both _serial_ and _parallel_ simulations

This test should take less than one minute. An example on how to visualise eSCAPE outputs in [**Paraview**](https://www.paraview.org/download/) is also provided.
