### 2018-09-25 Geodels/eSCAPE#3 @tristan-salles
Based on reviewer's comment I have updated the wiki page and also the model website:
+ [Installation](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC)
