### 2018-09-25 Geodels/eSCAPE#4 @tristan-salles
Based on reviewer's comment, the following section has been added to the [wiki](https://github.com/Geodels/eSCAPE/wiki/Input-file) and model website page.

## Additional informations

The tutorials proposed in the [eSCAPE-demo](https://github.com/Geodels/eSCAPE-demo) repository present several examples to create the required `vtk` input files (used in the input files for the domain, the climate, and the uplift) for running **eSCAPE**.

These input files are defined on an irregular triangular grid (TIN) which is created using either the [pyGmsh](https://pypi.org/project/pygmsh/) or [stripy](https://pypi.org/project/stripy/) libraries (both install on the Docker container).

In case of spatial change in forcing conditions, the user needs to specify (using the `map` element) for each vertices of the TIN, a series of fields for each forcing (tectonic displacements and precipitation) at chosen time intervals.

In cases where the forcing conditions are uniform over the entire domain one can chose to define its values in the YAML input file directly using the `uniform` element.

Sea level position can be set at a given relative position using the `position` element or using a `curve` which is a file containing 2 columns (time and sea-level position). In this later case, sea level position at any timestep is obtained by linear interpolation between closest sea level times.

## Outputs & Paraview visualisation

The model outputs are located in the output folder (`dir` element) and consist of a time series file named `eSCAPE.xdmf` and 2 other folders (`h5` and `xmf`). The **XDMF** file is the main entry point for visualising the output and should be sufficient for most users.

The file can be opened with the [**Paraview**](https://www.paraview.org/download/) software.

A video is provided in the [eSCAPE-demo](https://github.com/Geodels/eSCAPE-demo) repository **test/data** folder and shows how to visualise **eSCAPE** output.

As shown in the video, after loading the file, we perform 2 operations:
+ first we use the `wrap by scalar` filter to create a 2D representation of the surface
+ then we define a `contour` line corresponding to the sea-level position
