### 2018-09-25 Geodels/eSCAPE#7 @tristan-salles
Based on reviewer's comment the following paragraph has been added to the input file description in the wiki and model website:

***

Model-domain edge boundary conditions are set using the `bc` element in the YAML input file and the following options are available:

1. `flat`: edges elevations are set to the elevations of the closest non-edge vertices
2. `fixed`: edges elevations remain at the same position during the model run
3. `slope`: edges elevations are defined based on the closest non-edge vertices average slope
