### 2018-09-25 Geodels/eSCAPE#3 @kbarnhart
The updates to the documentation look great @tristan-salles ! I'll try them out on the HPC I have access to and report back if I run into any problems.
