### 2018-09-25 Geodels/eSCAPE#5 @kbarnhart
In looking at eSCAPE-demo more extensively, I see that that this is discussed there. You state:

> The code parallelisation relies on petsc4py and scalability tests are still on-going. For now on, we have seen some good performance up to 64 CPUs using a mesh with more than 6 millions vertices. Our goal is to be able to run models with more than 10 millions nodes and over several hundreds of CPUs.

This is very useful and I'd recommend stating this in the main eSCAPE repo instead.
