### 2018-09-25 Geodels/eSCAPE#2 @kbarnhart
@tristan-salles this is great.

I'd recommend you explicitly state that this test exists in the `readme.md` for the main eSCAPE repo.

Something like revising the sentence

> A set of examples are available in the eSCAPE-demo repository.

to

> A set of examples are available in the eSCAPE-demo repository. These examples include a small test that can be used to verify that the eSCAPE installation is working as expected.

There is also a very minor typo in the added text.

> The **test** proposes tow comparisons related to:

should be

> The **test** proposes two comparisons related to:
