### 2018-09-25 Geodels/eSCAPE#3 @kbarnhart
A minor comment:   `pandas` is listed with pip install but not listed under standard packages.
