### 2018-09-25 Geodels/eSCAPE#3 @kbarnhart
Comments related to the `petsc4py` installation

First, a minor one:  I think the statement `(see next section if you need to build petsc locally...)` should refer to petsc4py instead of petsc.

Second, my HPC platform already has `petsc4py` installed, the environment variable `$PETSC_DIR` is already set, and when I try and install the package I get the following message:

```
[barnhark@shas0136 barnhark]$ pip install petsc4py --user
Requirement already satisfied: petsc4py in /curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages
Requirement already satisfied: numpy in /curc/sw/python/2.7.11/lib/python2.7/site-packages (from petsc4py)
```

Thus I'd recommend revising the two sections that have to do with petsc to say something like the following:

> eSCAPE requires the PETSc library and petsc4py package. Here we will not describe compiling the PETSc library as it is often a default module available on HPC platforms.

> Depending on your particular HPC platform `petsc4py` may already be installed, a standard `pip install` will work, or you may need to compile `petsc4py` from scratch

Below this introduction, you can then recycle the text you've created. I'd recommend including an example of how a user would know that they already have `petsc4py` installed.

and a final minor comment about `petsc4py`. I'd change the two statement `export PETSC_DIR=...` to acknowledge that other HPC may have a different location for the `$PETSC_DIR`. Perhaps something like:  `export PETSC_DIR=\path\to\petsc\location`
