### 2018-09-25 Geodels/eSCAPE#5 @tristan-salles
I have added a diagram in the  eSCAPE-demo  README file regarding performance and updated the main eSCAPE repo
