### 2018-09-25 Geodels/eSCAPE#2 @tristan-salles
It has been added to the README file and the typo has been fixed! Thanks!
