### 2018-09-25 Geodels/eSCAPE#3 @tristan-salles
Now regarding the installation issue. It looks like you are using the intel compilers, did you install `fillit` and `eSCAPE` using the following commands:
```bash
export F90=ifort
git clone https://github.com/Geodels/fillit.git
cd fillit
python setup.py install --user

git clone https://github.com/Geodels/eSCAPE.git
cd eSCAPE
python setup.py install --user
```
with first `export F90=ifort` specified.
