### 2018-09-25 Geodels/eSCAPE#3 @kbarnhart
@tristan-salles

It looks like I forgot to include the lines that deal with installing `eSCAPE` and `fillit`. Sorry about that. I did execute `export F9=ifort` first.

I just checked the `ifort` version and it is older than the one you list. Don't know if that may be a problem.
```
[barnhark@shas0137 barnhark]$ ifort --version
ifort (IFORT) 17.0.4 20170411
Copyright (C) 1985-2017 Intel Corporation.  All rights reserved.
```

Based on a bit more debugging I think I have one issue related to the fortran compiler and one issue related to `petsc4py`.

Looking more closely at the terminal log that prints when I compile `fillit` and `eSCAPE` I see that I'm getting some warnings associated with not finding fortran compilers.  I assume I shouldn't be getting these fortran compiler warnings.

```
[barnhark@shas0137 barnhark]$ module purge
[barnhark@shas0137 barnhark]$ ml intel/17.4  impi/17.3
[barnhark@shas0137 barnhark]$ ml petsc/3.8.0
[barnhark@shas0137 barnhark]$ ml hdf5
[barnhark@shas0137 barnhark]$ export F90=ifort
[barnhark@shas0137 barnhark]$ cd fillit
[barnhark@shas0137 fillit]$ python setup.py install --user
make: Entering directory `/projects/barnhark/fillit/src'
ifort -shared -O2 -fPIC -c queues.f90
make: Leaving directory `/projects/barnhark/fillit/src'
/curc/sw/python/2.7.11/lib/python2.7/distutils/dist.py:267: UserWarning: Unknown distribution option: 'long_description_content_type'
  warnings.warn(msg)
running install
running bdist_egg
running egg_info
running build_src
build_src
building extension "fillit._fillBarnes" sources
f2py options: []
  adding 'build/src.linux-x86_64-2.7/fortranobject.c' to sources.
  adding 'build/src.linux-x86_64-2.7' to include_dirs.
building extension "fillit._fillWei" sources
f2py options: []
  adding 'build/src.linux-x86_64-2.7/fortranobject.c' to sources.
  adding 'build/src.linux-x86_64-2.7' to include_dirs.
building extension "fillit._fillZhou" sources
f2py options: []
  adding 'build/src.linux-x86_64-2.7/fortranobject.c' to sources.
  adding 'build/src.linux-x86_64-2.7' to include_dirs.
building data_files sources
build_src: building npy-pkg config files
writing fillit.egg-info/PKG-INFO
writing top-level names to fillit.egg-info/top_level.txt
writing dependency_links to fillit.egg-info/dependency_links.txt
reading manifest template 'MANIFEST.in'
warning: no files found matching 'fillit/Notebooks'
writing manifest file 'fillit.egg-info/SOURCES.txt'
installing library code to build/bdist.linux-x86_64/egg
running install_lib
running build_py
running build_ext
customize UnixCCompiler
customize UnixCCompiler using build_ext
customize Gnu95FCompiler
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/ifort
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/xild
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/xiar
Found executable /curc/sw/gcc/5.4.0/bin/gfortran
customize IntelFCompiler
customize LaheyFCompiler
Could not locate executable lf95
lahey: no Fortran 77 compiler found
lahey: no Fortran 77 compiler found
customize PGroupFCompiler
Could not locate executable pgfortran
pg: no Fortran 77 compiler found
pg: no Fortran 77 compiler found
customize AbsoftFCompiler
Could not locate executable f77
absoft: no Fortran 77 compiler found
absoft: no Fortran 77 compiler found
customize NAGFCompiler
Found executable /usr/bin/f95
customize VastFCompiler
customize CompaqFCompiler
Could not locate executable fort
compaq: no Fortran 77 compiler found
compaq: no Fortran 77 compiler found
customize IntelItaniumFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler using build_ext
creating build/bdist.linux-x86_64/egg
creating build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/queuepit.py -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/_fillWei.so -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/pitZhou.py -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/pitBarnes.py -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/pitWei.py -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/_fillZhou.so -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/__init__.py -> build/bdist.linux-x86_64/egg/fillit
copying build/lib.linux-x86_64-2.7/fillit/_fillBarnes.so -> build/bdist.linux-x86_64/egg/fillit
byte-compiling build/bdist.linux-x86_64/egg/fillit/queuepit.py to queuepit.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/pitZhou.py to pitZhou.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/pitBarnes.py to pitBarnes.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/pitWei.py to pitWei.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/__init__.py to __init__.pyc
creating stub loader for fillit/_fillBarnes.so
creating stub loader for fillit/_fillWei.so
creating stub loader for fillit/_fillZhou.so
byte-compiling build/bdist.linux-x86_64/egg/fillit/_fillBarnes.py to _fillBarnes.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/_fillWei.py to _fillWei.pyc
byte-compiling build/bdist.linux-x86_64/egg/fillit/_fillZhou.py to _fillZhou.pyc
installing package data to build/bdist.linux-x86_64/egg
running install_data
copying src/queues.o -> build/bdist.linux-x86_64/egg/fillit
copying src/queues.mod -> build/bdist.linux-x86_64/egg/fillit
running install_clib
customize UnixCCompiler
creating build/bdist.linux-x86_64/egg/EGG-INFO
copying fillit.egg-info/PKG-INFO -> build/bdist.linux-x86_64/egg/EGG-INFO
copying fillit.egg-info/SOURCES.txt -> build/bdist.linux-x86_64/egg/EGG-INFO
copying fillit.egg-info/dependency_links.txt -> build/bdist.linux-x86_64/egg/EGG-INFO
copying fillit.egg-info/top_level.txt -> build/bdist.linux-x86_64/egg/EGG-INFO
writing build/bdist.linux-x86_64/egg/EGG-INFO/native_libs.txt
zip_safe flag not set; analyzing archive contents...
creating 'dist/fillit-0.0.1-py2.7-linux-x86_64.egg' and adding 'build/bdist.linux-x86_64/egg' to it
removing 'build/bdist.linux-x86_64/egg' (and everything under it)
Processing fillit-0.0.1-py2.7-linux-x86_64.egg
Removing /home/barnhark/.local/lib/python2.7/site-packages/fillit-0.0.1-py2.7-linux-x86_64.egg
Copying fillit-0.0.1-py2.7-linux-x86_64.egg to /home/barnhark/.local/lib/python2.7/site-packages
fillit 0.0.1 is already the active version in easy-install.pth

Installed /home/barnhark/.local/lib/python2.7/site-packages/fillit-0.0.1-py2.7-linux-x86_64.egg
Processing dependencies for fillit==0.0.1
Finished processing dependencies for fillit==0.0.1
[barnhark@shas0137 fillit]$ cd ../eSCAPE
[barnhark@shas0137 eSCAPE]$ python setup.py install --user
running install
running build
running config_cc
unifing config_cc, config, build_clib, build_ext, build commands --compiler options
running config_fc
unifing config_fc, config, build_clib, build_ext, build commands --fcompiler options
running build_src
build_src
building extension "eSCAPE._fortran" sources
f2py options: []
  adding 'build/src.linux-x86_64-2.7/fortranobject.c' to sources.
  adding 'build/src.linux-x86_64-2.7' to include_dirs.
build_src: building npy-pkg config files
running build_py
running build_ext
customize UnixCCompiler
customize UnixCCompiler using build_ext
customize Gnu95FCompiler
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/ifort
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/xild
Found executable /curc/sw/intel/17.4/compilers_and_libraries_2017.4.196/linux/bin/intel64/xiar
Found executable /curc/sw/gcc/5.4.0/bin/gfortran
customize IntelFCompiler
customize LaheyFCompiler
Could not locate executable lf95
lahey: no Fortran 77 compiler found
lahey: no Fortran 77 compiler found
customize PGroupFCompiler
Could not locate executable pgfortran
pg: no Fortran 77 compiler found
pg: no Fortran 77 compiler found
customize AbsoftFCompiler
Could not locate executable f77
absoft: no Fortran 77 compiler found
absoft: no Fortran 77 compiler found
customize NAGFCompiler
Found executable /usr/bin/f95
customize VastFCompiler
customize CompaqFCompiler
Could not locate executable fort
compaq: no Fortran 77 compiler found
compaq: no Fortran 77 compiler found
customize IntelItaniumFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler using build_ext
running install_lib
running install_egg_info
Removing /home/barnhark/.local/lib/python2.7/site-packages/eSCAPE-0.1-py2.7.egg-info
Writing /home/barnhark/.local/lib/python2.7/site-packages/eSCAPE-0.1-py2.7.egg-info
running install_clib
customize UnixCCompiler
```

The second issue relates to `petsc4py`. I think that recommending that a user test that list of package imports would be great as it identified that my import issues start with `petsc4py`.

```python
>>> from petsc4py import PETSc
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/PETSc.py", line 6, in <module>
    PETSc = ImportPETSc(ARCH)
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/lib/__init__.py", line 29, in ImportPETSc
    return Import('petsc4py', 'PETSc', path, arch)
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/lib/__init__.py", line 64, in Import
    with f: module = imp.load_module(fullname, f, fn, info)
ImportError: /curc/sw/intel/17.4/mkl/lib/intel64/libmkl_sequential.so: undefined symbol: mkl_sparse_d_csr__g_n_syrk_notr_row_struct_i8
```

I'm getting my `petsc4py` package as part of a module that CU Research computing has built, so I'll ask them for help fixing that issue.

A very simple test of my `petsc4py` is as follows (and fails)

```
[barnhark@shas0137 barnhark]$ module purge
[barnhark@shas0137 barnhark]$ ml intel/17.4  impi/17.3
[barnhark@shas0137 barnhark]$ ml petsc/3.8.0
[barnhark@shas0137 barnhark]$ python
Python 2.7.11 (default, Mar 22 2016, 01:42:54)
[GCC Intel(R) C++ gcc 4.8 mode] on linux2
>>> import petsc4py
>>> petsc4py.__version__
'3.8.1'
>>> from petsc4py import PETSc
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/PETSc.py", line 6, in <module>
    PETSc = ImportPETSc(ARCH)
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/lib/__init__.py", line 29, in ImportPETSc
    return Import('petsc4py', 'PETSc', path, arch)
  File "/curc/sw/petsc/3.8.0/impi/17.3/intel/17.4/petsc4py/lib/python2.7/site-packages/petsc4py/lib/__init__.py", line 64, in Import
    with f: module = imp.load_module(fullname, f, fn, info)
ImportError: /curc/sw/intel/17.4/mkl/lib/intel64/libmkl_sequential.so: undefined symbol: mkl_sparse_d_csr__g_n_syrk_notr_row_struct_i8
```
