### 2018-09-26 Geodels/eSCAPE#3 @tristan-salles
Hi Katy,
The warnings are fine as the compilation is looking for potential compilers, below is the log for my compilation on USyD HPC:

```bash
running install
running build
running config_cc
unifing config_cc, config, build_clib, build_ext, build commands --compiler options
running config_fc
unifing config_fc, config, build_clib, build_ext, build commands --fcompiler options
running build_src
build_src
building extension "eSCAPE._fortran" sources
f2py options: []
  adding 'build/src.linux-x86_64-2.7/build/src.linux-x86_64-2.7/fortran/fortranobject.c' to sources.
  adding 'build/src.linux-x86_64-2.7/build/src.linux-x86_64-2.7/fortran' to include_dirs.
build_src: building npy-pkg config files
running build_py
running build_ext
customize UnixCCompiler
C compiler: icc -pthread -DNDEBUG -g -O3 -Wall -Wstrict-prototypes -I/usr/local/hdf5/1.8.14/include -I/usr/local/petsc-intel-mpi/3.9.3/include -I/usr/local/python/2.7.15-intel/include -I/usr/local/zlib/1.2.8/include -I/usr/local/cgal/4.9/include -I/usr/local/boost/1.64.0/include -I/usr/local/bzip2/1.0.6/include -I/usr/local/cairo/1.14.10/include -I/usr/local/ncurses/5.9/include -I/usr/local/libcroco/0.6.8/include -I/usr/local/openjpeg/2.1.0/include -I/usr/local/libcurl/7.40.0/include -I/usr/local/poppler/0.31.0/include -I/usr/local/libffi/3.2.1/include -I/usr/local/libspectre/0.2.7/include -I/usr/local/ghostscript/9.15/include -I/usr/local/perl/5.20.1/include -I/usr/local/glib/2.34.0/include -I/usr/local/harfbuzz/0.9.38/include -I/usr/local/libxml2/2.9.2/include -I/usr/local/pango/1.36.8/include -I/usr/local/pixman/0.32.6/include -I/usr/local/readline/6.3/include -I/usr/local/freetype/2.5.5/include -I/usr/local/pcre2/10.00/include -I/usr/local/libjpeg/6b/include -I/usr/local/fontconfig/2.11.92/include -I/usr/local/libtiff/4.0.3/include -I/usr/local/libpng/1.6.16/include -I/usr/local/openssl/1.1.0h/include -I/usr/local/gcc/4.9.3/include -I/usr/local/mpfr/3.1.4/include -I/usr/local/gmp/5.1.3/include -I/usr/local/hdf5/1.8.14/include -I/usr/local/petsc-intel-mpi/3.9.3/include -I/usr/local/python/2.7.15-intel/include -I/usr/local/zlib/1.2.8/include -I/usr/local/cgal/4.9/include -I/usr/local/boost/1.64.0/include -I/usr/local/bzip2/1.0.6/include -I/usr/local/cairo/1.14.10/include -I/usr/local/ncurses/5.9/include -I/usr/local/libcroco/0.6.8/include -I/usr/local/openjpeg/2.1.0/include -I/usr/local/libcurl/7.40.0/include -I/usr/local/poppler/0.31.0/include -I/usr/local/libffi/3.2.1/include -I/usr/local/libspectre/0.2.7/include -I/usr/local/ghostscript/9.15/include -I/usr/local/perl/5.20.1/include -I/usr/local/glib/2.34.0/include -I/usr/local/harfbuzz/0.9.38/include -I/usr/local/libxml2/2.9.2/include -I/usr/local/pango/1.36.8/include -I/usr/local/pixman/0.32.6/include -I/usr/local/readline/6.3/include -I/usr/local/freetype/2.5.5/include -I/usr/local/pcre2/10.00/include -I/usr/local/libjpeg/6b/include -I/usr/local/fontconfig/2.11.92/include -I/usr/local/libtiff/4.0.3/include -I/usr/local/libpng/1.6.16/include -I/usr/local/openssl/1.1.0h/include -I/usr/local/gcc/4.9.3/include -I/usr/local/mpc/1.0.3/include -I/usr/local/mpfr/3.1.4/include -I/usr/local/gmp/5.1.3/include -fPIC

creating /tmp/tmpYrARDS/tmp
creating /tmp/tmpYrARDS/tmp/tmpYrARDS
compile options: '-MMD -MF /tmp/tmpYrARDS/file.c.d -c'
icc: /tmp/tmpYrARDS/file.c
customize UnixCCompiler using build_ext
get_default_fcompiler: matching types: '['gnu95', 'intel', 'lahey', 'pg', 'absoft', 'nag', 'vast', 'compaq', 'intele', 'intelem', 'gnu', 'g95', 'pathf95', 'nagfor']'
customize Gnu95FCompiler
Found executable /usr/local/intel/18.1/bin/ifort
Found executable /usr/local/gcc/4.9.3/bin/gfortran
customize IntelFCompiler
customize LaheyFCompiler
Could not locate executable lf95
lahey: no Fortran 77 compiler found
lahey: no Fortran 77 compiler found
customize PGroupFCompiler
Could not locate executable pgfortran
pg: no Fortran 77 compiler found
pg: no Fortran 77 compiler found
customize AbsoftFCompiler
Could not locate executable f77
absoft: no Fortran 77 compiler found
absoft: no Fortran 77 compiler found
customize NAGFCompiler
Found executable /usr/bin/f95
customize VastFCompiler
customize CompaqFCompiler
Could not locate executable fort
compaq: no Fortran 77 compiler found
compaq: no Fortran 77 compiler found
customize IntelItaniumFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler
customize IntelEM64TFCompiler using build_ext
running install_lib
running install_egg_info
Removing /home/tsalles/.local/lib/python2.7/site-packages/eSCAPE-0.1-py2.7.egg-info
Writing /home/tsalles/.local/lib/python2.7/site-packages/eSCAPE-0.1-py2.7.egg-info
running install_clib
customize UnixCCompiler
C compiler: icc -pthread -DNDEBUG -g -O3 -Wall -Wstrict-prototypes -I/usr/local/hdf5/1.8.14/include -I/usr/local/petsc-intel-mpi/3.9.3/include -I/usr/local/python/2.7.15-intel/include -I/usr/local/zlib/1.2.8/include -I/usr/local/cgal/4.9/include -I/usr/local/boost/1.64.0/include -I/usr/local/bzip2/1.0.6/include -I/usr/local/cairo/1.14.10/include -I/usr/local/ncurses/5.9/include -I/usr/local/libcroco/0.6.8/include -I/usr/local/openjpeg/2.1.0/include -I/usr/local/libcurl/7.40.0/include -I/usr/local/poppler/0.31.0/include -I/usr/local/libffi/3.2.1/include -I/usr/local/libspectre/0.2.7/include -I/usr/local/ghostscript/9.15/include -I/usr/local/perl/5.20.1/include -I/usr/local/glib/2.34.0/include -I/usr/local/harfbuzz/0.9.38/include -I/usr/local/libxml2/2.9.2/include -I/usr/local/pango/1.36.8/include -I/usr/local/pixman/0.32.6/include -I/usr/local/readline/6.3/include -I/usr/local/freetype/2.5.5/include -I/usr/local/pcre2/10.00/include -I/usr/local/libjpeg/6b/include -I/usr/local/fontconfig/2.11.92/include -I/usr/local/libtiff/4.0.3/include -I/usr/local/libpng/1.6.16/include -I/usr/local/openssl/1.1.0h/include -I/usr/local/gcc/4.9.3/include -I/usr/local/mpfr/3.1.4/include -I/usr/local/gmp/5.1.3/include -I/usr/local/hdf5/1.8.14/include -I/usr/local/petsc-intel-mpi/3.9.3/include -I/usr/local/python/2.7.15-intel/include -I/usr/local/zlib/1.2.8/include -I/usr/local/cgal/4.9/include -I/usr/local/boost/1.64.0/include -I/usr/local/bzip2/1.0.6/include -I/usr/local/cairo/1.14.10/include -I/usr/local/ncurses/5.9/include -I/usr/local/libcroco/0.6.8/include -I/usr/local/openjpeg/2.1.0/include -I/usr/local/libcurl/7.40.0/include -I/usr/local/poppler/0.31.0/include -I/usr/local/libffi/3.2.1/include -I/usr/local/libspectre/0.2.7/include -I/usr/local/ghostscript/9.15/include -I/usr/local/perl/5.20.1/include -I/usr/local/glib/2.34.0/include -I/usr/local/harfbuzz/0.9.38/include -I/usr/local/libxml2/2.9.2/include -I/usr/local/pango/1.36.8/include -I/usr/local/pixman/0.32.6/include -I/usr/local/readline/6.3/include -I/usr/local/freetype/2.5.5/include -I/usr/local/pcre2/10.00/include -I/usr/local/libjpeg/6b/include -I/usr/local/fontconfig/2.11.92/include -I/usr/local/libtiff/4.0.3/include -I/usr/local/libpng/1.6.16/include -I/usr/local/openssl/1.1.0h/include -I/usr/local/gcc/4.9.3/include -I/usr/local/mpc/1.0.3/include -I/usr/local/mpfr/3.1.4/include -I/usr/local/gmp/5.1.3/include -fPIC

creating /tmp/tmpoA0Nl3/tmp
creating /tmp/tmpoA0Nl3/tmp/tmpoA0Nl3
compile options: '-MMD -MF /tmp/tmpoA0Nl3/file.c.d -c'
icc: /tmp/tmpoA0Nl3/file.c
```

Yes I will add the import test in the wiki installation
