### 2018-09-26 #964 @sgrieve
I would like to thank the author for this contribution, it has been a very enjoyable experience to explore this software and I am pleased to recommend publication, pending the minor issues highlighted below being resolved.

# Version
I noticed that there is a version 1.0.1 on GitHub, but the submission corresponds to version 1.0.0. It would be good to clarify this, so that the current stable version is the one that is reviewed.

# Installation
In the first instance I explored the code using Docker. Everything worked out of the box, and running the test notebook yielded the correct results with no surprises or problems. I would however, suggest that a little more detail in the main `readme.md` file on how to use Kitematic and Docker may be useful, as I imagine that this is the most likely method of install for new users, and those not familiar with Docker may get held up on this step.

One issue I noticed was that there is no `requirements.txt` file at the base of the repo, containing all of the python dependencies and their version requirements as listed on the HPC install page on the wiki, this would save people a bit of hassle during the HPC install process.

# Functionality

Working my way through the extensive tutorials and examples showed me the range of applications for eSCAPE, and I am happy that the functional claims made in the documentation and the paper are valid. More broadly I believe that eSCAPE will be a valuable tool for researchers across the geosciences, and that it is providing a novel solution to the common problem of scale in the modeling of surface processes.

# Performance

I have been unable to run this software on an HPC system (due to my current lack of access to one) so have only been able to explore the performance of eSCAPE on my MacBook. Where performance claims have been made, they have been in line with my experience of running eSCAPE. As I have tested larger and larger model runs, I have been pleasantly surprised by the compute times and believe that the performance of this model is in line with other similar model's performance, if not exceeding them.

# Documentation

Going through the documentation I am impressed at the detail and quality of what has been produced. Most of the issues I would have flagged have already been resolved following the issues created by @kbarnhart.

One further issue I noticed was that in the homepage of the API documentation, there is a table listing submodules (`mesher`, `flow`, `pit`, `tools`). This table has no associated information about what these submodules are/what they do. Similarly, on the contents pane of the API docs there are links to these submodules (`eSCAPE.mesher`, `eSCAPE.flow`, `eSCAPE.pit`, `eSCAPE.tools`), but these links do not contain any information. I realize that Sphinx can be a pain for stuff like this, but if possible it would be good to either populate these pages if there is relevant information to go into them, or remove the links to avoid confusion.

Some other minor issues:

- The `About` section of the wiki just contains some images, and no information.
- The `Contact` section of the wiki links back to the wiki homepage, and provides no contact information
- In the `runModel` notebook for the Australia tutorial, the ParaView command `Warp by Scalar` is called `Wrap by Scalar`. (I think this issue might be repeated in some of the other notebooks.)

# Tests

The test notebook that is provided works well to ensure that the code is running as expected, and provides detailed information to users to begin debugging if things do not go as expected. In an ideal world there would be some kind of CI set up to run automated tests, but I appreciate the difficulties in doing this. I know that some UK research councils have a national Jenkins instance which supports automated testing and I don't know if that is something you can explore with your funders?

# Paper

The paper clearly outlines the motivation for the development of eSCAPE, and the key theoretical underpinnings of the work. It reads clearly and follows the general pattern for JOSS papers.

Overall I am delighted to have had the opportunity to review this paper, and have been thoroughly impressed by the overall quality of the software, documentation and tutorials. I hope to continue using the software in the future! :rocket:
