### 2018-09-26 #964 @kbarnhart
I thought I'd give a quick update on the progress of my review. As noted by @sgrieve I've made most of my comments in the form of issues. @tristan-salles has dealt with most of these impressively fast. This has been very helpful as I was able to attempt HPC installation based on the revised HPC instructions instead of the original ones.

I am testing the HPC installation instructions so between the two reviewers we will have tested all installation options. As of yesterday evening am hung up on a step related to the `petsc4py` installation done by my institutions Research Computing staff. @tristan-salles and I have been iterating on improvements to the HPC installation instructions on [eSCAPE Issue #3](https://github.com/Geodels/eSCAPE/issues/3). I don't expect that resolving my HPC issues will take too long. Once I've been able to confirm HPC install I'll be able finalize my review.

I anticipate that I will recommend acceptance with minor revisions (many of which @tristan-salles has already completed).
