### 2018-09-27 6998c92
update docstrings for Sphinx generation
