+++ doc/conf.py
@@ -26,7 +26,7 @@ author = u'Tristan Salles'
 # The short X.Y version
 version = u''
 # The full version, including alpha/beta/rc tags
-release = u'v.1.0.1'
+release = u'v.1.0.0'
 
 
 # -- General configuration ---------------------------------------------------
