+++ eSCAPE/flow/__init__.py
@@ -16,4 +16,8 @@
 # along with eSCAPE.  If not, see <http://www.gnu.org/licenses/>.
 ###
 
+
+"""
+Surface process module for hillslope and stream erosion/deposition.
+"""
 from .surfprocplex import SPMesh
