### 2018-09-27 #964 @tristan-salles
A massive thank you to @kbarnhart and @sgrieve for their quick and super valuable inputs, comments & feedbacks!

### Comments from @sgrieve

I think I have pretty much addressed all the comments. In details I have:
+ fixed the version to 1.0.0 which is now consistent with the one from the paper
+ added a `requirements.txt` file
+ updated the wiki with a step by step guide in regards to installing and using Docker and Kitematic ([see](https://github.com/Geodels/eSCAPE/wiki/Using-Docker))
+ updated the API to populated the submodule table and removing unnecessary submodules links ([see](https://escape-api.github.io/index.html))
+ updated the [About section of the wiki](https://github.com/Geodels/eSCAPE/wiki/About)
+ updated the [Contact section of the wiki](https://github.com/Geodels/eSCAPE/wiki/Contact)
+ fixed the typo in the demo python notebooks (`warp by scalar` to `wrap by scalar`)

Regarding the testing issue, I have added a `test installation` example in the `eSCAPE-demo` repository that provides two comparisons related to:
+ the _expected values_ that you should obtained if your installation is successful
+ runtime for both _serial_ and _parallel_ simulations
This test should take less than 1 minute. An example on how to visualise **eSCAPE** outputs in [**Paraview**](https://www.paraview.org/download/) is also provided.

### Comments from @kbarnhart

I think I have answered most of the issues and comments there (see the history of closed issues above). There is still a remaining open one: **Improve eSCAPE HPCC installation instructions Geodels/eSCAPE#3**.
