### 2018-09-27 #964 @sgrieve
Thanks for responding to my comments so quickly @tristan-salles!

@kthyng I'm satisfied that all of my comments have been addressed, so pending the resolution of the HPC install issues that @kbarnhart has raised I'm happy to formally accept!
