### 2018-09-27 Geodels/eSCAPE#3 @kbarnhart
I wanted to give a quick update on my end here.

I think that the remaining compilation errors are related to my HPC platform (my use of `petsc4py` uncovered that it hadn't been compiled correctly against MKL). Thats been fixed. I still haven't successfully compiled `fillit` or `eSCAPE`. The errors are related to `__stack_chk_guard` which CU RC staff has told me is related to Intel switching the version of GCC used on the backend (and that they've seen these errors come up before related to compiling python packages).

They are helping me work through this and I expect to have things ironed out Friday/tomorrow or early next week. Ideally I'd like to successfully compile and run `eSCAPE` on an HPC before I finalize my review. However, if I encounter additional delays associated with libraries/compilers available on the HPC platform I am working on I won't let this delay my finalizing a review in a reasonable timeline.
