+++ fortran/functionsmodule.h
@@ -10,8 +10,6 @@
 #define  initDiffCoeff   initdiffcoeff
 #define  getMaxEro   getmaxero
 #define  getDiffElev   getdiffelev
-#define  distributeHeight   distributeheight
-#define  distributeVolume   distributevolume
 #define  MFDreceivers   mfdreceivers
 #define  defineTIN   definetin
 
@@ -27,8 +25,6 @@
 #define  initDiffCoeff_   initdiffcoeff_
 #define  getMaxEro_   getmaxero_
 #define  getDiffElev_   getdiffelev_
-#define  distributeHeight_   distributeheight_
-#define  distributeVolume_   distributevolume_
 #define  MFDreceivers_   mfdreceivers_
 #define  defineTIN_   definetin_
 
@@ -44,8 +40,6 @@
 #define  _initDiffCoeff   _initdiffcoeff
 #define  _getMaxEro   _getmaxero
 #define  _getDiffElev  _getdiffelev
-#define  _distributeHeight   _distributeheight
-#define  _distributeVolume   _distributevolume
 #define  _MFDreceivers  _mfdreceivers
 #define  _defineTIN  _definetin
 
@@ -61,7 +55,5 @@
 #define  _initDiffCoeff_   _initdiffcoeff_
 #define  _getMaxEro_  _getmaxero_
 #define  _getDiffElev_  _getdiffelev_
-#define  _distributeHeight_   _distributeheight_
-#define  _distributeVolume_   _distributevolume_
 #define  _MFDreceivers_  _mfdreceivers_
 #define  _defineTIN_   _definetin_
