+++ eSCAPE/flow/surfprocplex.py
@@ -28,8 +28,6 @@ import warnings;warnings.simplefilter('ignore')
 from eSCAPE._fortran import setHillslopeCoeff
 from eSCAPE._fortran import initDiffCoeff
 from eSCAPE._fortran import MFDreceivers
-from eSCAPE._fortran import distributeHeight
-from eSCAPE._fortran import distributeVolume
 from eSCAPE._fortran import getMaxEro
 from eSCAPE._fortran import getDiffElev
 
@@ -169,9 +167,6 @@ class SPMesh(object):
         self.wghtVal[self.pitID,:] = 0.
         # Account for marine regions
         self.seaID = np.where(hArrayLocal<self.sealevel)[0]
-        self.rcvID[self.seaID,:] = np.tile(self.seaID,(self.flowDir,1)).T
-        self.distRcv[self.seaID,:] = 0.
-        self.wghtVal[self.seaID,:] = 0.
         del hArrayLocal
 
     	if MPIrank == 0 and self.verbose:
@@ -190,6 +185,7 @@ class SPMesh(object):
         # Build drainage area matrix
         if self.rainFlag:
             WAMat = self.iMat.copy()
+            WeightMat = self._matrix_build_diag(np.zeros(self.npoints))
             for k in range(0, self.flowDir):
                 tmpMat = self._matrix_build()
                 data = -self.wghtVal[:,k].copy()
@@ -200,12 +196,13 @@ class SPMesh(object):
                 tmpMat.setValuesLocalCSR(indptr, self.rcvID[:,k].astype(PETSc.IntType), data,
                                                                 PETSc.InsertMode.INSERT_VALUES)
                 tmpMat.assemblyEnd()
+                WeightMat += tmpMat
                 WAMat += tmpMat
                 tmpMat.destroy()
 
             # Solve flow accumulation
             WAtrans = WAMat.transpose()
-            self.WAtrans = WAtrans.copy()
+            self.WeightMat = WeightMat.copy()
             self.dm.localToGlobal(self.bL, self.bG, 1)
             self.dm.globalToLocal(self.bG, self.bL, 1)
             if self.tNow == self.tStart:
@@ -214,6 +211,7 @@ class SPMesh(object):
                 self._solve_KSP(True,WAtrans, self.bG, self.drainArea)
             WAMat.destroy()
             WAtrans.destroy()
+            WeightMat.destroy()
         else:
             self.drainArea.set(0.)
         self.dm.globalToLocal(self.drainArea, self.drainAreaLocal, 1)
@@ -260,35 +258,64 @@ class SPMesh(object):
             # Solve flow accumulation
             self._solve_KSP(True, WHMat, self.hOld, self.hGlobal)
             WHMat.destroy()
+            del Kcoeff
 
+        # Nullify underwater erosion
+        self.dm.globalToLocal(self.hGlobal, self.hLocal, 1)
+        hL0 = self.hLocal.getArray().copy()
+        hL0[self.seaID] = self.hOldArray[self.seaID]
+        self.hLocal.setArray(hL0)
+        self.dm.localToGlobal(self.hLocal, self.hGlobal, 1)
+
+        # Update erosion thicknesses
         self.stepED.waxpy(-1.0,self.hOld,self.hGlobal)
         self.cumED.axpy(1.,self.stepED)
         self.dm.globalToLocal(self.cumED, self.cumEDLocal, 1)
         self.dm.localToGlobal(self.cumEDLocal, self.cumED, 1)
 
-        # Update elevation locally due to aerial erosion
-        hL0 = self.hLocal.getArray().copy()
-        self.dm.globalToLocal(self.hGlobal, self.hLocal, 1)
-
         # Get sediment load value for given time step
[…]
