### 2018-10-01 #964 @kthyng
I see that @sgrieve is ready to accept, and I see the issue referenced above that is in progress on @kbarnhart's end. Thanks for the great progress and work, @tristan-salles, @sgrieve, and @kbarnhart!
