### 2018-10-02 #964 @kbarnhart
Alright. I'm completing my review now @kthyng .

## Barnhart eSCAPE review

I'm happy to recommend acceptance contingent on minor revision.

This is an excellent contribution and I look forward to having applications that
require using it.

My primary concern is with HPC installation instructions. As I expect that many
major research-grade applications (and/or coupling with additional models) will
require a user to compile the code, I think that ironing these instructions out
is important.

## Explanation for few checkboxes I didn't check.

I don't think any of these things should prevent this paper from moving forward.

I didn't check '**Installation**: Does installation proceed as outlined in the
documentation?' as I was not able to surmount some of my HPC platform issues and
thus I never successfully installed this on an HPC. Ultimately the issue seems
to be related to the details of the Fortran compiler used in the backend of the
Intel compiler version that the staff at the University of Colorado Research
computing used to build PETSc.
@tristan-salles is in no way at fault in this and was very helpful throughout my attempt to compile. Based on my attempts I think
that the installation instructions are **almost** sufficient (see comments
below).

I also didn't check '**Installation instructions**: Is there a clearly-stated
list of dependencies? Ideally these should be handled with an automated package
management solution.' I think that both the `requirements.txt` file and the HPC
installation instructions need to be revised (more details below).

I didn't check '**Automated tests**: Are there automated tests or manual steps
described so that the function of the software can be verified?' because there
are no automated tests. However there are tests that meet the [JOSS "good"
requirements](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). I think there are reasonable reasons that there are no automated
tests (e.g. need a platform with PETSc, etc, to run the tests).

## Recommended revision to to installation instructions.

I think that there are two items that relate to installation that need to be
addressed. The first relates to the `requirements.txt` and the second relates
to the HPC install instructions wiki page.

## requirements.txt

I think that the `requirements.txt` file needs to be further revised in order
to include the minimum version numbers (where applicable). I also think that
having the `fillit` and `meshplex` package in the `requirements.txt` file is a
bit awkward as these are not packages that are distributed on standard package
distributors (conda, pip). Further the `meshplex` package is a modified version
forked by @tristan-salles. I see no problem with these as dependencies, but
putting these in a `requirements.txt` file means that a command like.

```
pip install -r requirements.txt
```

will fail because these packages are not findable.

## HPC install instructions wiki

I think that the [HPC installation page on the wiki](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC)
needs to be further revised to not be so specific to one HPC platform. I also
think that text related to the fortran compiler used is confusing.

### Specificity regarding platform

Sections like [this one on example dependencies](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC#dependencies-installation-example)
show loading specific modules like petsc built with gcc and open MPI. I would
recommend removing these lines and replacing with regular (not code block) text
that says something like:

> Before you begin compiling eSCAPE and dependencies, make sure you have set up
your compute environment to include python, OpenMPI, a supported compiler, and
PETSc.

Now that a `requirements .txt` file exists, all the pip install commands can be
replaced by a single call

```
pip install -r requirements.txt
```

Combining all the code needed to install on the Artemis HPC at USyD is useful,
but perhaps should go on another page called "Example installation scripts used on specific HPC platforms"
Once I get my platform working I would happily submit my script there. This would help
draw the distinction for a user that there are *general* requirements, and that there are
*specific* examples that worked on a particular platform.

### Fortran compiler

There are two places in the examples where the fortran compiler F90 is assigned
to something else (ifort in one place and gfortran in another). I will admit to
not knowing much about different fortran compilers (or different compilers in general...)
and thus these instructions are confusing to me. I might guess that I should be
using gfortran if I used gcc instead of the intel compiler, and ifort if I used
the intel compiler. I would recommend only writing

```
export F90=something
```

in one place on the HPC instructions and providing a bit of context as to why
someone should use gfortran or ifort in this situation.

### Minor comments about the example
The [example PBS script](https://github.com/Geodels/eSCAPE/wiki/Installation-on-HPC#example-of-pbs-script)
assumes that a user has downloaded the `eSCAPE-demo` repository and dealt with all
of its requirements. It should be revised to include information about downloading
the demo repository. I also think that it would be more appropriate on another wiki
page.
