### 2018-10-02 42b29aa
add bedrock/sediment cover differentiation
