+++ eSCAPE/flow/surfprocplex.py
@@ -221,6 +221,163 @@ class SPMesh(object):
 
         return
 
+    def _getErosionRate(self,Hsoil):
+        """
+        Compute sediment and bedrock erosion rates.
+        """
+
+        Kcoeff = self.drainAreaLocal.getArray()
+        Kbr = np.power(Kcoeff,self.mbr)*self.Kbr
+        Ksed = np.power(Kcoeff,self.msed)*self.Ksed
+        EbedMat = self.iMat.copy()
+        EsedMat = self.iMat.copy()
+
+        # Define erosion coefficients
+        for k in range(0, self.flowDir):
+
+            indptr = np.arange(0, self.npoints+1, dtype=PETSc.IntType)
+            nodes = indptr[:-1]
+
+            # Bedrock erosion processes computation
+            if self.Kbr > 0.:
+                data = np.divide(Kbr, self.distRcv[:,k], out=np.zeros_like(Kcoeff),
+                                            where=self.distRcv[:,k]!=0)
+                tmpMat = self._matrix_build()
+                data = np.multiply(data,self.wghtVal[:,k])
+                data[self.rcvID[:,k].astype(PETSc.IntType)==nodes] = 0.0
+                tmpMat.assemblyBegin()
+                tmpMat.setValuesLocalCSR(indptr, self.rcvID[:,k].astype(PETSc.IntType), data,
+                                                                PETSc.InsertMode.INSERT_VALUES)
+                tmpMat.assemblyEnd()
+                EbedMat += tmpMat
+                Mdiag = self._matrix_build_diag(data)
+                EbedMat -= Mdiag
+                tmpMat.destroy()
+                Mdiag.destroy()
+                del data
+
+            # Sediment erosion processes computation
+            if self.Ksed > 0.:
+                data = np.divide(Ksed, self.distRcv[:,k], out=np.zeros_like(Kcoeff),
+                                            where=self.distRcv[:,k]!=0)
+                tmpMat = self._matrix_build()
+                data = np.multiply(data,self.wghtVal[:,k])
+                data[self.rcvID[:,k].astype(PETSc.IntType)==nodes] = 0.0
+                tmpMat.assemblyBegin()
+                tmpMat.setValuesLocalCSR(indptr, self.rcvID[:,k].astype(PETSc.IntType), data,
+                                                                PETSc.InsertMode.INSERT_VALUES)
+                tmpMat.assemblyEnd()
+                EsedMat += tmpMat
+                Mdiag = self._matrix_build_diag(data)
+                EsedMat -= Mdiag
+                tmpMat.destroy()
+                Mdiag.destroy()
+                del data
+
+        # Solve bedrock erosion rate
+        self._solve_KSP(True, EbedMat, self.hOld, self.vGlob)
+        EbedMat.destroy()
+        self.stepED.waxpy(-1.0,self.hOld,self.vGlob)
+        E = self.stepED.getArray().copy()
+        Ecrit = np.divide(E, self.crit_br, out=np.zeros_like(E),
+                               where=self.crit_br!=0)
+        E -= self.crit_br*(1.0-np.exp(-Ecrit))
+        E[E<0.] = 0.
+        E = np.multiply(E,np.exp(-Hsoil/self.Hstar))
+        E = np.multiply(E,1.0-self.frac_fine)
+        self.Eb.setArray(E)
+        self.dm.globalToLocal(self.Eb, self.EbLocal, 1)
[…]
