+++ requirements.txt
@@ -1,16 +1,9 @@
-setuptools
-gfortran 
-python
-cython
-petsc
-numpy
-scipy
-pandas
-mpi4py
-petsc4py
-hdf5
-h5py
-ruamel.yaml
-meshio
-meshplex
-fillit
+numpy==1.15.1
+scipy==1.1.0
+Cython==0.28.5
+mpi4py==3.0.0
+petsc4py==3.8.1
+meshio==2.2.5
+h5py==2.8.0
+pandas==0.17.1
+ruamel.yaml==0.15.64
