### 2018-10-03 Geodels/eSCAPE#3 @kbarnhart
@tristan-salles thanks for these improvements! I'm closing this issue.
