### 2018-10-03 #964 @kbarnhart
@kthyng a quick note that I've now closed all the issues I opened in the eSCAPE repository.
