+++ eSCAPE/flow/surfprocplex.py
@@ -162,10 +162,11 @@ class SPMesh(object):
         self.rcvID, self.slpRcv, self.distRcv, self.wghtVal = MFDreceivers(self.flowDir, self.inIDs, hArrayLocal)
         # Account for pit regions
         self.pitID = np.where(self.slpRcv[:,0]<=0.)[0]
-        self.rcvID[self.pitID,:] = np.tile(self.pitID,  (self.flowDir,1)).T
+        self.rcvID[self.pitID,:] = np.tile(self.pitID, (self.flowDir,1)).T
         self.distRcv[self.pitID,:] = 0.
         self.wghtVal[self.pitID,:] = 0.
         # Account for marine regions
+        # self.seaID = hArrayLocal<self.sealevel
         self.seaID = np.where(hArrayLocal<self.sealevel)[0]
         del hArrayLocal
 
@@ -297,13 +298,14 @@ class SPMesh(object):
         EsedMat.destroy()
         self.stepED.waxpy(-1.0,self.hOld,self.vGlob)
         E = self.stepED.getArray().copy()
-        # self.ESed = Es.copy()
         Ecrit = np.divide(E, self.crit_sed, out=np.zeros_like(E),
                                where=self.crit_sed!=0)
         E -= self.crit_sed*(1.0-np.exp(-Ecrit))
         E[E<0.] = 0.
-        E = np.multiply(E,1.0-np.exp(-Hsoil/self.Hstar))
-        E[E*self.dt>Hsoil] = Hsoil/self.dt
+        E = np.multiply(E,self.dt*(1.0-np.exp(-Hsoil/self.Hstar)))
+        ids = E>Hsoil
+        E[ids] = Hsoil[ids]
+        E = np.divide(E,self.dt)
         E = np.multiply(E,1.0-self.phi)
         self.Es.setArray(E)
         self.dm.globalToLocal(self.Es, self.EsLocal, 1)
@@ -312,7 +314,7 @@ class SPMesh(object):
         self.EsLocal.setArray(E)
         self.dm.localToGlobal(self.EsLocal, self.Es, 1)
 
-        del E, Ecrit
+        del E, Ecrit, ids
         del Kcoeff, Kbr, Ksed
 
         return
@@ -353,7 +355,6 @@ class SPMesh(object):
         Qs = self.vland*tmpQ
         Qs[self.seaID] = self.vsea*tmpQ[self.seaID]
         depo = np.divide(Qs, Qw, out=np.zeros_like(Qw), where=Qw!=0)
-
         # Update volume of sediment to distribute
         Qs = np.multiply(depo,self.FVmesh_area)
         tmpQ[self.pitID] += Qs[self.pitID]
@@ -365,18 +366,20 @@ class SPMesh(object):
         Qs[self.idGBounds] = 0.
 
         # Set marine deposition volume
-        depo = depo.fill(0.)
+        depo.fill(0.)
         tmpQ = Qs*self.dt/(1.0-self.phi)
         depo[self.seaID] = tmpQ[self.seaID]
         self.vLoc.setArray(depo)
-        # self.vLoc.setArray(tmpQ)
 
         # Set elevation change due to deposition inland
-        depo = np.divide(tmpQ, self.FVmesh_area, out=np.zeros_like(tmpQ), where=FVmesh_area!=0)
+        tmpQ[self.seaID] = 0.
+        depo = np.divide(tmpQ, self.FVmesh_area, out=np.zeros_like(tmpQ), where=self.FVmesh_area!=0)
+        self.vecL.setArray(depo)
+        self.dm.localToGlobal(self.vecL, self.vecG, 1)
 
-        del Qw, Qs, tmpQ
+        del Qw, Qs, tmpQ, depo
 
-        return depo
[…]
