### 2018-10-04 #964 @kthyng
@kbarnhart Has @tristan-salles completed or otherwise addressed the minor revisions you noted or are there any outstanding issues?
