### 2018-10-04 #964 @kbarnhart
Yes! Absolutely! No outstanding issues remain. I am happy to recommend
publication. Sorry if that wasn’t previously clear.

On Thu, Oct 4, 2018 at 3:37 PM Kristen Thyng <notifications@github.com>
wrote:

> @kbarnhart <https://github.com/kbarnhart> Has @tristan-salles
> <https://github.com/tristan-salles> completed or otherwise addressed the
> minor revisions you noted or are there any outstanding issues?
>
