### 2018-10-05 d6312d1
add initial soil thickness map capability
