### 2018-10-08 #964 @kthyng
Ok thanks @kbarnhart, just wanted to be sure.

@arfon This paper has been accepted by the reviewers (@kbarnhart and @sgrieve)!!!! Ready for publication!!
