### 2018-10-08 #964 @kthyng
@tristan-salles Please create an archive (on Zenodo, figshare, or other) and post the archive DOI in this issue.
