### 2018-10-09 #964 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
